Move old credentials to `Credentials_old` and new system to `Credentials`
