+++ python/Ganga/GPIDev/Credentials/AfsToken.py
