+++ python/Ganga/GPIDev/Credentials/CredentialStore.py
