+++ python/Ganga/GPIDev/Credentials/ICredentialInfo.py
