+++ python/Ganga/GPIDev/Credentials/ICredentialRequirement.py
