+++ python/Ganga/GPIDev/Credentials_old/ICredential.py
