+++ python/Ganga/Lib/LCG/Grid.py
@@ -3,7 +3,7 @@ import re
 import tempfile
 import datetime
 
-from Ganga.GPIDev.Credentials2 import credential_store
+from Ganga.GPIDev.Credentials import credential_store
 
 from Ganga.Utility.Config import getConfig
 from Ganga.Utility.logging import getLogger
