+++ python/Ganga/Lib/LCG/LCG.py
@@ -36,8 +36,8 @@ from Ganga.Utility.logic import implies
 from Ganga.GPIDev.Base.Proxy import isType, getName
 from Ganga.Utility.GridShell import getShell
 
-from Ganga.GPIDev.Credentials2 import require_credential, credential_store, needed_credentials
-from Ganga.GPIDev.Credentials2.VomsProxy import VomsProxy
+from Ganga.GPIDev.Credentials import require_credential, credential_store, needed_credentials
+from Ganga.GPIDev.Credentials.VomsProxy import VomsProxy
 
 from . import Grid
