+++ python/GangaDirac/Lib/Utilities/DiracUtilities.py
@@ -10,7 +10,7 @@ from Ganga.Utility.Config import getConfig
 from Ganga.Utility.logging import getLogger
 from Ganga.Core.exceptions import GangaException
 from Ganga.GPIDev.Base.Proxy import isType
-from Ganga.GPIDev.Credentials2 import credential_store
+from Ganga.GPIDev.Credentials import credential_store
 import Ganga.Utility.execute as gexecute
 
 logger = getLogger()
